Move compare to tools lib
